Modularized Model Styles and Actions
